+++ internal/posse/app.go
@@ -17,7 +17,7 @@ import (
 )
 
 const (
-	Version       = "0.4.0"
+	Version       = "0.5.0"
 	FallbackEmoji = "⚙️"
 )